refactor: ScheduleAddViewController와 ScheduleModifyViewController를 ScheduleEditViewController 하나로 합침 #68 - ScheduleEditCoordinatorType, ScheduleEditViewModelType 프로토콜 생성
