+++ Weekand/Weekand/Presentation/MainScene/ScheduleEdit/ViewModel/DefaultRepeatViewModel.swift
@@ -12,12 +12,12 @@ import RxCocoa
 
 class DefaultRepeatViewModel: ViewModelType {
     
-    weak var coordinator: ScheduleCoordinatorType?
+    weak var coordinator: ScheduleEditCoordinatorType?
     private var disposeBag = DisposeBag()
     
     var repeatType: ScheduleRepeatType?
     
-    init(coordinator: ScheduleCoordinatorType, repeatType: ScheduleRepeatType) {
+    init(coordinator: ScheduleEditCoordinatorType, repeatType: ScheduleRepeatType) {
         self.coordinator = coordinator
         self.repeatType = repeatType
     }
@@ -39,7 +39,7 @@ extension DefaultRepeatViewModel {
     @discardableResult
     func transform(input: Input) -> Output {
         input.cancelButtonDidTapEvent.subscribe(onNext: {
-            if let coordinator = self.coordinator as? ScheduleEditCoordinator {
+            if let coordinator = self.coordinator as? ScheduleAddCoordinator {
                 coordinator.finish()
             } else if let coordinator = self.coordinator as? ScheduleModifyCoordinator {
                 coordinator.finish()
@@ -53,7 +53,7 @@ extension DefaultRepeatViewModel {
             .withLatestFrom(selectRepeatEndDate)
             .subscribe(onNext: { [weak self] date, isRepeat in
                 let repeatEndDate = isRepeat ? date : nil
-                if let coordinator = self?.coordinator as? ScheduleEditCoordinator {
+                if let coordinator = self?.coordinator as? ScheduleAddCoordinator {
                     coordinator.sendRepeatTypeFromSheet(repeatType: self?.repeatType ?? .once, repeatEndDate: repeatEndDate)
                     coordinator.finish()
                 } else if let coordinator = self?.coordinator as? ScheduleModifyCoordinator {
