+++ Weekand/Weekand/Presentation/MainScene/ScheduleEdit/ViewModel/ScheduleEditViewModelType.swift
@@ -1,0 +1,10 @@
+//
+//  ScheduleViewModelType.swift
+//  Weekand
+//
+//  Created by 이호영 on 2022/07/27.
+//
+
+import Foundation
+
+protocol ScheduleEditViewModelType: AnyObject, ViewModelType { }
